Move validationSetsSuite to snapstate_install_test.go
